requirements.txt
ruff==0.12.0
mypy==1.17.0
delta-spark==4.0.0
pyarrow==21.0.0
pyspark==4.0.0